Initial commit.
We've got some functionality. Specifically all explicit dependencies
from an entry point. There is also a dot helper to render svg's in
potentially useful layouts. This is best used scripted on small
chunks of large codebases. Too many nodes and it is less than useful.
Batching of subsections is advised, but after deduping graph edges, it
will likely be less of a problem than before, but further testing
should be done.